pyproject.toml
[project]
name = "synthetic_ocr_data"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Marie Roald", email = "roald.marie@gmail.com"},
]
dependencies = [
    "pillow>=10.4.0",
    "fonttools>=4.53.1",
    "pandas>=2.2.2",
    "augraphy @ file:///home/mariero/augraphy",
    "opencv-python>=4.10.0.84",
    "colorspacious>=1.1.2",
    "datasets>=3.0.0",
    "lxml>=5.3.0",
    "tqdm>=4.66.5",
    "corpustools @ git+https://github.com/divvun/CorpusTools.git",
    "numpy>=2.0.2",
    "jupyter>=1.1.1",
    "jinja2>=3.1.4",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
dev = [
    "ruff>=0.6.3",
    "pytest>=8.3.2",
    "coverage>=7.6.1",
    "pytest-coverage>=0.0",
    "rich>=13.8.1",
    "pytest-cov>=5.0.0",
]
[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["I", "TCH"]

[tool.ruff.lint.isort]
known-first-party = ["synthetic_text_images"]

[tool.pytest.ini_options]
testpaths = ["tests", "src"]
addopts = "--doctest-modules --cov src --import-mode importlib"
# --doctest-modules runs doctests as well
# --cov src is for pytest-cov to measure coverage in the src directory (using coverage.py) 
# --import-mode importlib allows multiple tests files to have the same name
